sql: pull out multi-region finalization into its own struct
This thing was getting slightly unwieldy, so it makes sense to
encapsulate all the finalization logic in one place and a separate
file. Inspired by otan's draft PR #63459.

Release note: None
